requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20